Use separate timer per interception context when DatabaseLogFormatter is used without a context
Issue #22

In the most common use, the DatabaseLogFormatter is used for a single DbContext, which means that there can be no interleaving between starting and finishing an execution. However, if the formatter is being used for multiple contexts, then these calls can be interleaved, which means that each pair needs its own timing.

Since each call has its own interception context, this can be used to store the timer. We accepted a pull request to store this kind of data, but I realized that if there are multiple interceptors and each want to use this data slot, the result would be one clobbering the other. Therefore, I obsoleted that mechanism in favor of a mechanism that can be used by multiple interceptors.

The change required that Stopwatch be obsoleted. However, the existing functionality has been retained so that an in-place update or ignoring of the obsolete warning will not make working code break.